requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
playwright==1.56.0
beautifulsoup4==4.14.3
openai==1.54.0
reportlab==4.2.5
aiofiles==23.2.1
